Clear pip cache on github actions cache miss (#2)
It seems like sometimes the restore of the github actions cache fails and when it does it leaves the partially restored cache hanging around. In this case our build fails because pip doesn't install the packages we need.

This should clean up anything leftover after a cache miss so we avoid this case.