Upgrade to Pillow 9.0.0 (#167)
This upgrades to Pillow 9.0.0. This required updating one test image
for a failing test, as the rendered text evidently has ever-so-slightly
different spacing when rendered by the new Pillow. This also reduced
the tolerance for the RMS check in the affected test, because it was
set unnecessarily high due to a mistake in the code ("Authpr" →
"Author").