Refactor DRM scheme/content type handling
This introduces a new class, `FormatPriorities`, which is used to
filter and sort the delivery mechanisms available from a license
pool by DRM scheme and content type. This is used when producing
acquisition links for OPDS entries.

Currently, the OPDS feed generation code uses `FormatPriorities`
to filter mechanisms for feed entries, but no interface is
currently provided to actually configure the `FormatPriorities` class.
The `FormatPriorities` class is designed to give unconfigured behaviour
equivalent to the old (arbitrary) behaviour of iterating over delivery
mechanisms in database order.