Avoid accidentally storing unhashed passphrases
The existing code had several places where hashed and unhashed
passwords could be easily mixed up, because they were both just plain
strings. Additionally, the code had an apparent mistake where an
unhashed password would be generated and then stored in the database
as if it was a hashed password.

This introduces LCPHashedPassphrase and LCPUnhashedPassphrase classes
to distinguish between the two types, and adds a method to produce a
hashed passphrase from an unhashed passphrase. Type annotations were
added along with a few type assertions at key points. The reason for
the assertions is that we don't yet have static checking of types,
and the consequences for making a mistake are dire: Invalid data ends
up in the database.

Affects: https://www.notion.so/lyrasis/Invalid-LCP-hashed-passphrases-returned-for-some-library-cards-213c107b48fe492baa497e42f5e0d68e