Adjust "required" settings check
This refactors the "required" settings check to be slightly easier
to debug and to also fix an issue where a setting with a default
value of None would still be considered to be "required" due to the
presence of the "default" key.

This also adjusts a couple of the API tests to take into account that
the Overdrive configuration parameters now have unique prefixes.

Affects: https://www.notion.so/lyrasis/Migrate-Overdrive-integration-to-new-configuration-API-1f8d105bf6cd4c99a23bd6baead15d9e