Implement a mock web server for understandable tests
This introduces a new MockAPIServer test fixture that can be used
to verify requests and responses made to servers without having to
patch the internals of Python platform classes.

This also adds a test that verifies that requests are _not_ retried
on error codes.

Affects: https://www.notion.so/lyrasis/overdrive_new_title-import-fails-due-to-timeouts-and-500-errors-87ec0491e0f94a889a3ef37cc2f440a5