Only search for lanes within the current library
The endpoint that creates/edits lanes was calling get_one() without
passing in the ID of the library being edited. This resulted
in get_one() possibly returning lanes that are defined for other
libraries, and this would cause the check that determines if a display
name is already used to fail.

Affects: https://www.notion.so/lyrasis/Inability-to-create-a-custom-lane-with-the-same-name-in-each-library-on-the-same-CM-f549719021a54fb1be0ec1882cc42cfd#d9f576304c5c478aa2bdf9b11cd3a028