Make the block_reason check consistent.
This implements some missing logic to avoid performing checks, and to
set the block_reason field to NO_VALUE to be consistent with (most of)
the rest of the codebase. This also adds a check in the Patron class
to correctly handle a block_reason of `NO_VALUE`.

Affects: https://www.notion.so/lyrasis/Make-the-do-not-block-SIP-setting-actually-avoid-blocking-37fa459ca93044c4b5a44668de6d580e