Fix LCP prioritization
The previous "de-prioritize LCP content" feature was broken because the
configuration settings were exposed in a manner that allowed the Admin
UI to place string labels into the database instead of the enumeration
values that were intended to go in instead. This switches to a model
where we store boolean values, and includes a database migration to
fix up any databases that were affected.

Affects: https://www.notion.so/lyrasis/De-prioritize-LCP-switch-is-broken-c9b295a78c1142a1922ed2d7e3143836