Slightly improve the exception behaviour
This adjusts the retry logic such that individual books are retried up to a fixed
maximum (currently 10) number of retries, and if a book is still failing after the
retries, it is ignored and the import is continued. The import will be marked as
having failed, as a whole, but it will run to the end.

Affects: https://www.notion.so/lyrasis/Brooklyn-PL-overdrive_new_title-fails-updating-work-coverage-records-7c42a0c4278d4c5c8ad91c986a4a90b2