Add a library-name parameter
This adjusts the import script such that it now requires a library-name
parameter to import lists to a specific library. This is in contrast
to the old behaviour where all lists were examined and their library
names were extracted from OPDS feeds.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc