Attempt to use ID values from URIs instead of id elements
This adjusts the code to take book ID and ID types from the links
inside OPDS feeds, instead of depending on the value of the "id"
element (which may not be usable as a work ID on the target CM). This
also adjusts the code to remove any percent encoding that may be in
use on the identifiers.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc