Check the correct identifiers on the import side
This updates the code to use the identifiers from the importing CM's
OPDS feed entry links. It improves the error message to avoid implying
that the title is the problem when the problem might actually be with
the identifier.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc