Attempt, once again, to get imports working
This adjusts the JSON schema and code to carry the full raw identifier
of books, the identifier types of books, and the identifier values
of books. All three items are needed separately to call the various
CM endpoints. New tests have been added to check that the various
identifier types are captured correctly.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc