Attempt to use ID values from URIs instead of id elements (#442)
* Attempt to use ID values from URIs instead of id elements

This adjusts the code to take book ID and ID types from the links
inside OPDS feeds, instead of depending on the value of the "id"
element (which may not be usable as a work ID on the target CM). This
also adjusts the code to remove any percent encoding that may be in
use on the identifiers.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc

* Correct a mypy issue

* Check the correct identifiers on the import side

This updates the code to use the identifiers from the importing CM's
OPDS feed entry links. It improves the error message to avoid implying
that the title is the problem when the problem might actually be with
the identifier.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc

* Attempt, once again, to get imports working

This adjusts the JSON schema and code to carry the full raw identifier
of books, the identifier types of books, and the identifier values
of books. All three items are needed separately to call the various
CM endpoints. New tests have been added to check that the various
identifier types are captured correctly.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc