Attempt to work around a failure to update coverage records (#394)
* Attempt to work around a failure to update coverage records

This catches the StaleDataError exception that might be raised
by a long-running Overdrive import when something else happens to
modify overlapping work coverage records at the same time the import
runs. This is, at best, a workaround, and may not actually fix the
problem.

Affects: https://www.notion.so/lyrasis/Brooklyn-PL-overdrive_new_title-fails-updating-work-coverage-records-7c42a0c4278d4c5c8ad91c986a4a90b2

* Slightly improve the exception behaviour

This adjusts the retry logic such that individual books are retried up to a fixed
maximum (currently 10) number of retries, and if a book is still failing after the
retries, it is ignored and the import is continued. The import will be marked as
having failed, as a whole, but it will run to the end.

Affects: https://www.notion.so/lyrasis/Brooklyn-PL-overdrive_new_title-fails-updating-work-coverage-records-7c42a0c4278d4c5c8ad91c986a4a90b2

* Make retry message slightly more descriptive

* Switch maximum retries to 3

* Fix test failures