Correct a possible mistake in import id checking (#454)
It appears that the code was comparing the wrong id elements, leading
to an import failure and a misleading error message. This should fix
that. This also introduces more debug logging, so we can see what's
gone wrong on the next inevitable failure.

Affects: https://www.notion.so/lyrasis/Migrate-discrete-entries-in-client-library-lists-from-old-CM-to-new-CM-1b78b13b10234ee59830c343ef4862dc