Millenium patron authentication integration should support alternative patron identifiers (#463)
* Allow using different Millenium patron fields as barcodes

This adds an option to the admin UI to specify a field that will be
used as a barcode instead of the default 'pb' field. Some systems
evidently use fields such as the username field ('pu') as the patron
identifier. This changes the code such that one other field can be
treated as a patron identifier, and the effect is additive. Specifically:

  * If a 'pb' field exists, it is added as a potential identifier.
  * If the configuration specifies that another field `x` should
    be treated as a patron identifier, it is also added as a potential
    identifier.
  * The field `x` will _also_ be processed exactly as it would have
    been before this change, so for example if the configuration
    specifies that the 'pu' field will be used as a patron identifier,
    it will _also_ be processed and used as a patron username.

Affects: https://www.notion.so/lyrasis/Millenium-patron-authentication-integration-should-support-alternative-patron-identifiers-f1213e709eae4294aeccdb82a6abea94

* Add missing test file

* Only treat the configured field as the patron identifier

This changes the behaviour from the earlier commit such that _only_
the field given in the configuration will be treated as a patron
identifier.

Affects: https://www.notion.so/lyrasis/Millenium-patron-authentication-integration-should-support-alternative-patron-identifiers-f1213e709eae4294aeccdb82a6abea94

* Fix comments!

* Adjust comment wording