Add an initial database test fixture
This introduces a database test fixture that allows for:

  * Obtaining session-scoped application initialization
  * Obtaining session-scoped database initialization
  * Obtaining function-scoped transactions

Fixtures are layered such that dependencies between initialization
code are explicit; if a fixture needs a database, then it must ask for
a database. If the database needs the application to be initialized
first, then it must ask for it, and so on.

One test case is included to check that database initialization works
when asking for a transaction.

Affects: https://www.notion.so/lyrasis/Refactor-circulation-tests-to-use-pytest-fixtures-4f2603de236746aeaab123dcf12aa445