Send a GET request to Elasticsearch
This sends a GET request to Elasticsearch in tests instead of waiting
two seconds in the hope that the index has been updated. Unfortunately,
this doesn't seem to save any time, but it is at least more correct.

Affects: https://www.notion.so/lyrasis/Refactor-circulation-tests-to-use-pytest-fixtures-4f2603de236746aeaab123dcf12aa445