Detailed version endpoint (#765)
This PR adds a new endpoint version.json to the circulation manager, that replaces the old healthcheck endpoint. This endpoint returns version information for the circulation manager instance, and for the admin UI. This is currently helpful, but will be even more helpful when we are doing more rapid deployments, so that we can check what version is deployed where.

The new endpoint returns something like this:
```json
{
  "admin_ui": {
    "package": "@thepalaceproject/circulation-admin",
    "version": "0.4.2"
  },
  "branch": "main",
  "commit": "00f33eb39163937cfe864537124a43058e216a57",
  "version": "1.0.1"
}
```

The version is derived from the git tag by Dunamai. If the commit is not a tagged version, it will create a version number based on the distance from the last tagged version and the commit hash. Something like 1.0.2-post.51+00f33eb39.