Fix UpdateLaneSizeScript causing performance issue
Add a new local flag to lanes, that allows a lane update to suppress
the listener that triggers an update to the configuration time stamp.
This lets the UpdateLaneSizeScript update all the lanes before changing
the timestamp.

This should resolve a performance problem we are seeing on our
production servers, especially the ones with lots of lane.

This is a quick fix, while we work through the more complicated solution
to the problem of removing the need for the configuration reload
mechanism at all.