Make passwords required (in most cases)
This updates the code in the Admin API controller that handles password
changes. Now, passwords are always required for new admins.

There are numerous issues with the original password checking code that
will require a lot of refactoring, but these are all outside of the scope
of the ticket associated with this work.

Affects: https://www.notion.so/lyrasis/Make-password-a-required-field-when-creating-an-account-on-the-CM-ab4313f97fc4465687dfbafc9f64dc1e