Reduce time to run API tests (#800)
Running the API tests locally, this branch runs the API tests in ~1/3 of the time. Taking about 10 minutes instead of 30 minutes.

I did some profiling to try to understand why initializing the circulation manager was taking so long in production. While doing that I found that a significant portion of the time spend running the API tests was repeatedly generating cryptographic keys.

This just mocks out the key generation, and replaces it with a constant static key while we are running our tests.