Remove 🔥 CDN Config (#785)
Currently we don't use the CDN configuration anywhere, and when we add a CDN in the future I'd like to take a different approach, instead of storing it in the database.

In doing some planning work around caching, I ran into some difficulty with how the CDN configuration gets a new DB session. Instead refactoring the CDN code to initialize differently, it seemed like a good time to remove this code instead.