Remove 🔥 unused core config (#786)
This is a follow up PR to #785. As part of that work I noticed how much cursed 💀 💣 code is buried in the core config class. This code seems like its leftover from before config settings were added to the database, and config used to be loaded from a file.

Most of the code in here is actually dead, and never called. And what is being called, was just returning defaults, since a config file was never loaded.

This PR refactors the small bits of the class that are left, and removes as much of the old code as possible.