Attempt to require passwords (again)
The original code to handle the creation or editing of admins had
a serious bug in that, even if the endpoint returned an error, the
code would still commit the admin that it might have created in
the process of trying to check if the admin existed. This refactors
the method into something much more explicit, where the three possible
code paths are kept distinct, and an explicit rollback is made if
anything other than success occurs.

Affects: https://www.notion.so/lyrasis/Make-password-required-when-changing-already-existing-admin-account-information-in-the-CM-36df0e10860547d4a6a693eea452f37d