Upgrade Flask and Werkzeug (#855)
Upgrade the version of Flask and Werkzeug we are using to the latest versions.

There are some outstanding CVE against Werkzeug that make it desirable to upgrade it. However once we start upgrading we quickly end up in a dependency hell 🔥 that ends in flask-sqlalchemy-session not supporting the most recent flask version.

Since flask-sqlalchemy-session hasn't seen an update since 2015, and is very small, only about 50 lines of code. We pull that code into core and apply some patches to get it working with the latest versions of flask and werkzeug.