Fix failing tests due to werkzeug escaping changes (#887)
Currently our tests are failing on the main branch, due to a conflict between that tests that started running in #835 and #855. This is caused by Werkzeug not escaping as many characters in URLS see issue here pallets/werkzeug#262. These URLs are still valid, but have less characters in them escaped.