Replace configuration settings for auth providers (#1063)
Creates two new tables:
- IntegrationConfiguration
- IntegrationLibraryConfiguration

These tables store information about external integrations, as well as their settings. These new tables replace the old ExternalIntegration and ConfigurationSettings tables.

The settings for the integrations are stored in a json column in the database, and serialized into a Pydantic dataclass. The dataclass handles all the validation of the input data submitted by the admin interface. The settings form is configured as part of the Pydantic dataclass using the field function on each form value.

This PR also replace the settings for all of the authentication providers with this new setup as a proof of concept. Future iterations will need to be done to move our other external integrations to the new settings model.