Migrate discovery services to integration config (PP-94) (#1321)
Convert the library registry to use IntegrationConfiguration rather then ExternalIntegration. The library registration data wasn't a great fit for LibraryIntegrationConfiguration, so that data is stored in its own table discovery_service_registrations.

As part of the conversion, I moved the files for the registry into a discovery module and fully type hinted them, so it was a bit easier to sort out what was happening. As part of the type hinting process, I added hints to core/util/problem_detail.py, which necessitated some changes in files unrelated directly to this PR.

See JIRA ticket: PP-94.
This is part of the larger CM Config Settings epic: PP-4.