Update OPDS2 token fulfillment using opaque patron identifier (PP-416) (#1374)
## Description
Instead of directly sending the patron username or external_identifier to the token auth endpoint, instead use the identifier_to_remote_service function to create an identifier to send to the remote service.

## Motivation and Context
In PP-416, OPDS2 with token auth is failing because the patron record doesn't have the required fields (either username or external_identifier). We've seen failures like this in the past as well. By generating an opaque identifier and sending that instead we make sure that we are able to use token auth with any valid patron auth, eliminating one source of errors, and helping to improve patron privacy.