Configure logging integrations via environment variables (PP-496) (#1421)
Update logging to be setup as part of the service container via environment variables, instead of being configured using configuration settings from the database.

This should fix the issue we have been seeing where CI hangs and eventually times out. Since the code that is causing the hang is configuring logging in the database. And is part of the work we need to do anyway to convert our settings fully away from integrationsettings stored in the database.