OPDS Importer Base Classes (PP-500) (#1428)
Refactor the OPDS importer class hierarchy so that we have some base classes, that the specific importers implement. There are two new base classes: BaseOPDSImporter and BaseODLImporter. These classes contain all of the shared functionality across OPDSImporter, OPDS2Importer, ODLImporter, ODL2Importer and OPDSForDistributorsImporter, so its a little easier to see what it overridden and when looking at the classes.

This also tried to make the tests for ODL and ODL2 a bit easier to understand by not having them inherit from each other and instead use a parameterized fixture that picks the correct API for the tests.

This refactor is setting up for PP-502 to replace external integrations with configurationsettings in the collection integrations.