Refactor CirculationAPI to remove unused parameters (PP-500) (#1437)
Remove the unused part and fulfill_part_url parameters from the fulfill call in CirculationAPI. Instead of passing internal_format into each API, let the APIs opt into this with a CirculationInternalFormatsMixin mixin class.

Removing internal_format makes sense because only a couple APIs actually use it, the others ignore it. The forthcoming OPDS and OPDS2 APIs will also ignore this, so it made sense to refactor it out.