Additional information for web process in JSONFormatter (PP-1148) (#1771)
When troubleshooting issues though cloudwatch logs on a busy server, its difficult to correlate particular requests made to the server to the log messages that are output, since many requests are coming in at once. I've come across this a couple times recently when troubleshooting.

This adds some additional keys to the output of JSONFormatter and to our UWSGI log output, so we can better correlate logs to requests.