Query optimization for the /admin/stats endpoint (PP-1311) (#1870)
Reduce the number of sql queries we are doing when hitting the stats endpoint by doing some of the aggregation on the database side.

Using a copy of the production CA database on my laptop, the /admin/stats endpoint takes about 30 - 40 seconds for a admin user to load and about 15 - 20 seconds for a normal non-admin user. This is without the database under load, I would guess the performance is worse in production.

After these changes for me locally the admin user takes about 3 seconds to load, and a non-admin user is under a second. The admin user is still slower then I would like, but 10x speedup is a good start here.