Remote patron activity sync celery task (PP-1164) (#1894)
This PR updates our Patron API that sync with remote providers to do the sync as an async celery task, rather then in the main web thread worker. To coordinate these patron activity sync tasks, I wired up redis as a service that we can use to store data, not just as a celery broker.

Redis is used to store the state of the patron activity sync and as a distributed lock, so that only one task is doing a sync at a time.