Update indexing task to use bulk indexing endpoint (PP-1382) (#1909)
This PR updates search indexing so that it is done in bulk via the open search bulk api endpoint. This should significantly speed up indexing. As works need indexing we add them to a redis set, this also has the helpful side effect of de-duplicating the works we need to index.

Once a minute we queue up a task removes the works from the redis set, and queues up another task that bulk indexes them into elastic search.

We also add a new TaskLock class, that creates a distributed redis lock to make sure only a single task of a particular type is running at a time. This task lock is used in the new search_indexing task.