Delegate Fulfillment instead of using FulfillmentInfo (PP-1641) (#2035)
Refactor FulfillmentInfo class which just contained information about a fufillment, but had multiple methods to override how fulfillment was done, to be a new class called Fulfillment which the loans controller delegates fulfillment to.

This lets each API control how fulfillment is done, and reduces the amount of if / then we need to have in the loan controller. We define several types of Fulfillment, corresponding to the logic that used to exist in the loans controller, and update the API classes to use them.

The goal of this PR is the have fulfillment for each API act the same way before and after, just refactoring to use the new structure.