Don't proxy the full response headers (PP-1713) (#2055)
This solves an issue we were seeing where the Content-Encoding header is
set on the request, but since Python requests automatically decodes the
transfer encoding, we were mistakenly forwarding on the Content-Encoding
header, without encoded data.

Note: This commit is a HOTFIX and differs from #2055 due to a
merge conflict. It should only be used until a full release
is ready.