Update ODL hold calculation logic (PP-1728) (#2117)
Completely re-work how we are calculating the hold queue for OPDS2 + ODL items.

All the work to calculate the holds is moved out to live in a Celery task. When a hold is invoked through the API the hold is always placed in the queue, and the queue is later recalculated by the celery task.