Fix get_one_or_create IntegrityError handler (PP-1838) (#2134)
Fix an issue with the Exception handler for an IntegrityError in the get_one_or_create function.

If one of the parameters for this function is a sqlalchemy class, then the logging function can cause a PendingRollbackError, since the rollback is happening after the log message.

Update the function to use db.begin_nested() as a context manager, and write a test for this case.

I added type hints to a couple related functions while I was debugging this before i narrowed in on exactly the cause. These hints are included in this PR. They aren't really related to this fix, but they seem harmless to come in with this one. I can break them out to separate PR if desired.