"Could not place hold" error with multiple licenses (PP-1902) (#2173)
When we have multiple licenses, it was possible for us to get out of sync with the provider, so we would try to place a hold, when a different license could have been tried instead. This updates the checkout to try all available licenses, in priority order, before resorting to a hold.