Filter on library when checking lane name unique (PP-1967) (#2190)
This PR does a small bit of refactoring to DRY the check we use for lane name uniqueness and fixes a bug where the check for lane name uniqueness

Previously one of the checks was missing a filter on the lanes library. Making it so that you couldn't create a lane with the same name as any other lane in the CM, rather then any other lane with the same library.