Refactor our analytics providers (#2323)
## Description

This PR refactors our analytics providers to:
- Create a common abstract base class `AnalyticsProvider`
- Create a pydantic model for the json data that we output to s3
`AnalyticsEventData`
- This model contains all the information needed to publish an analytics
event, disconnected from the DB session, so we are able to save it
across transaction boundaries without worrying about stale objects.
- Update our analytics providers to consume `AnalyticsEventData`
- The local analytics provider and the s3 analytics provider both did
overlapping serialization of the analytics data, they both now rely on
the data provided by `AnalyticsEventData`
- Update opds_odl celery tasks to use the new `AnalyticsEventData`
structure, instead of reattaching DB objects across the session.

## Motivation and Context

Did this refactoring as part of the work I was doing in PP-1470, as some
of the reapers need to be able to persist event data across sessions
similar to the opds_odl celery tasks

## How Has This Been Tested?

- Running tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.