Convert database_reaper to celery (PP-1470) (#2324)
## Description

Convert all of the `ReaperMonitor` scripts to be Celery Tasks:
- `CredentialReaper`
- `PatronRecordReaper`
- `WorkReaper`
- `CollectionReaper`
- `MeasurementReaper`
- `LoanReaper`
- `HoldReaper`
- `IdlingAnnotationReaper`

Updates some of our database foreign key relationships so its easier to
use a database level delete cascade to clean up old records in the
database, rather then relying on the ORM to do the cascade.

## Motivation and Context

Working through the epic to convert all of our scripts into tasks.

JIRA: PP-1470