Remove storing patron neighborhood / location 🔥 (PP-2210) (#2326)
## Description

Remove the patron neighborhood / location storage.

## Motivation and Context

When converting the database reaper scripts to celery in PP-1470, I
didn’t convert the “scrubber“ scripts that removed patron location
stored in the database after a certain period of time.

Instead of putting that work into a feature we’re not using, I pulled it
out for now. If / when we need it we can look at generalizing it and
adding it back in.