[PP-2233] Fix deadlocks by ensuring only one book is updated per tran… (#2333)
…saction.

## Description

The PR attempts to solve the problem of deadlocks occurring between the
same type of tasks. The deadlocks appear to occur because some CMs have
many Axis collections which are pulling the same data. As a result they
were frequently updating the same batches of IDs at the same time. Since
the simultaneous updates can touch the same database objects - such as
editions, this can result in a deadlock in the following way:

1.  task 1 opens transaction
2. task 2 opens transaction
3. task 1 updates editions A, B, C
4. task 2 updates editions A, B, C
5. task  1  tries to commit but fails due to dead lock

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2233