Create ability to make delivery mechanisms unavailable (#2346)
## Description

This PR introduces functionality to track available delivery mechanisms
through a new flag in the `LicensePoolDeliveryMechanism` model. The main
changes:

1. Added an `available` flag to `LicensePoolDeliveryMechanism` that
defaults to `True`
2. Modified the `available_delivery_mechanisms` relationship in
`LicensePool` to filter by the `available` flag
3. Maintained the original `delivery_mechanisms` relationship that
returns all mechanisms regardless of availability

## Motivation and Context

This work is the basis for the changes in PP-2218 for Overdrive. This
change allows us to mark delivery mechanisms as unavailable when we try
to fulfill them and they turn out not to be available.

I figured I'd break this one out into its own PR, since the changes were
getting big already, and there is a lot of changes necessary to
implement the Overdrive specific part of this.

## How Has This Been Tested?

- Added new tests
- Ran tests in CI

## Checklist

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.