Return exception when Overdrive book in unfulfillable format (PP-2218) (#2381)
## Description

Its possible for some Overdrive titles to not be available in a format
that the mobile apps can work with. We don't know this until after the
checkout is complete.

In this case we return the item and raise a `CannotLoan` exception, so
the mobile apps can display an error message.

## Motivation and Context

This is another case found in testing with the mobile apps.

## How Has This Been Tested?

- Added new fixture file
- Added tests for this case

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.