Bugfix for http verbose logging (PP-2319) (#2384)
## Description

Fix bug in our HTTP class that was causing an exception when making a
request with `verbose=True`. It was causing the failures in PP-2319.

It was causing this exception to occur:
```
Log message could not be formatted. Exception: TypeError('not all arguments converted during string formatting')
```

## Motivation and Context

Was caused when I converted the logging statement to use a f-string in
https://github.com/ThePalaceProject/circulation/pull/2366. After
converting it, I forgot to remove the args to the log function 🤦🏻

## How Has This Been Tested?

- Added a test for this case, since the verbose option wasn't being
tested at all. This would have caught the issue.

## Checklist

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.