Move the LicenseStatus enum into OPDS package (#2394)
## Description

Move the `LicenseStatus` enum into OPDS package

## Motivation and Context

This enum is really part of our OPDS model, so it should be in the OPDS
package. Having it there also means that importing the OPDS models,
doesn't mean importing all of our sqlalchemy models. This is helpful for
this PR: https://github.com/ThePalaceProject/palace-tools/pull/107

## How Has This Been Tested?

- Ran unit tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.