Better logging OPDS2 publications fail validation (PP-2184) (#2390)
## Description

Make sure we are logging relevant details when an item in our OPDS2 feed
fails validation.

## Motivation and Context

I was trying to provide a log query with the relevant details in
PP-2184, but we were not adding this information to the logs. After this
change should be able to provide a query that will help find problematic
items.

## How Has This Been Tested?

- Tested locally with feed

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.