Fix Overdrive loans that cannot be viewed in mobile clients (PP-2372) (#2402)
## Description

This PR does two related things. They are in separate commits to make
code review a bit easier.

- Handle the case where an overdrive ebook is available in multiple
formats, none of which we support. We've seen this case in our real CMs
where the formats available after a checkout are: `{'ebook-overdrive',
'ebook-kindle'}`. Before this change, we were failing to detect that as
a case that the mobile apps couldn't view.
- Rename the `supported_formats` property on the `Checkout` model to
`available_formats`, since the previous name was a bit misleading, and
contributed to this bug.

## Motivation and Context

Issue came in via support. See ticket: PP-2372. 
Certain OverDrive titles on iOS staying in constant state of
"Requesting" after clicking Download. Currently the mobile clients
cannot handle a book that has no way to fulfill after checkout. We are
handling this in the backend, but were failing to detect this case.

## How Has This Been Tested?

- Added new test case
- Tested locally with problematic title

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.